+++ CHANGES.rst
@@ -5,6 +5,7 @@ Changelog
 ------------------
 
 - Include toggleSelect for Plone 5.
+- Disable columns in searchreplaceform view.
   [Gagaro]